+++ test/test_helper.exs
@@ -30,8 +30,8 @@ extra_exclude =
   if ch_version >= "25" do
     []
   else
-    # Time and JSON types are not supported in ClickHouse < 25
-    [:time, :json]
+    # Time, Variant, and JSON types are not supported in older ClickHouse versions we have in the CI
+    [:time, :variant, :json]
   end
 
 ExUnit.start(exclude: [:slow | extra_exclude])